requirements.txt
blinker==1.6.3
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
et-xmlfile==1.1.0
flask==3.0.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.1
openpyxl==3.1.2
pandas==2.1.1
Pillow==10.1.0
pymongo==4.5.0
pypng==0.20220715.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
typing-extensions==4.8.0
tzdata==2023.3
werkzeug==3.0.0
zipp==3.17.0
